deletecomment is up

+++ resources/views/comment/post.blade.php
@@ -10,11 +10,23 @@
     @foreach ($comment as $item)
     <div class="well">
         {{$item->comment}}
-        @if (Auth::user()->matrik_id == $item->student_id)
-        {{Form::open( ['action' => ['CommentController@deletecomment',$item->id], 'method' => 'delete' ])}}
-        {{Form::submit('Delete', ['class' => 'btn btn-info pull-right'] )}}
-        {{Form::close()}} 
+
+        @if (Auth::guard('web')->check())
+            @if (Auth::guard('web')->user()->matrik_id == $item->commenter_id )
+                {{Form::open( ['action' => ['CommentController@deletecomment',$item->id], 'method' => 'delete' ])}}
+                {{Form::submit('Delete', ['class' => 'btn btn-info pull-right'] )}}
+                {{Form::close()}} 
+            @endif
+
+        @elseif(Auth::guard('supervisor')->check())
+            @if (Auth::guard('supervisor')->user()->super_matrik_id == $item->commenter_id )
+                {{Form::open( ['action' => ['CommentController@deletecomment',$item->id], 'method' => 'delete' ])}}
+                {{Form::submit('Delete', ['class' => 'btn btn-info pull-right'] )}}
+                {{Form::close()}} 
+            @endif
         @endif
+        
+
 
         
     </div>
